--- app/_graphql/queries/album/album-works.ts
@@ -1,14 +1,0 @@
-import { partialWorkFieldsFragment } from "@/_graphql/fragments/partial-work-fields"
-import { graphql } from "gql.tada"
-
-export const albumWorksQuery = graphql(
-  `query AlbumWorks($albumId: ID!, $offset: Int!, $limit: Int!) {
-    album(id: $albumId) {
-      id
-      works(offset: $offset, limit: $limit) {
-        ...PartialWorkFields
-      }
-    }
-  }`,
-  [partialWorkFieldsFragment],
-)
