--- app/_graphql/queries/album/albums-count.ts
@@ -1,7 +1,0 @@
-import { graphql } from "gql.tada"
-
-export const albumsCountQuery = graphql(
-  `query AlbumsCount($where: AlbumsWhereInput) {
-    albumsCount(where: $where)
-  }`,
-)
